5.8 - ADDED view and add functions to index.js 6 - unmodularized everything!!!
