+++ lib/deleteFunctions.js
@@ -58,3 +58,18 @@
 //     deleteRole,
 //     deleteEmployee
 // }
+
+      // // DELETE from department WHERE ID = ___;
+      // else if (answers.menu === "Remove a department.") {
+      //   deleteDepartment()
+      // }
+
+      // // DELETE from role WHERE ID = ___;
+      // else if (answers.menu === "Remove a role.") {
+      //   deleteRole()
+      // }
+
+      // // DELETE from employee WHERE ID = ___;
+      // else if (answers.menu === "Remove an employee.") {
+      //   deleteEmployee()
+      // }
